refactor: remove /components/forms folder
