--- src/components/forms/commons/FormBase/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./FormBase";
-export * from "./FormBase";
