--- src/components/forms/commons/index.ts
@@ -1,5 +1,0 @@
-import FormBase from "./FormBase";
-import type { FormBaseProps } from "./FormBase";
-
-export { FormBase };
-export type { FormBaseProps };
