--- src/components/forms/index.ts
@@ -1,1 +1,0 @@
-export * from "./commons";
